fix: column types for foreign key constraints

+++ database/migrations/2020_02_19_183418_create_bon_exchange_table.php
@@ -28,7 +28,7 @@ class CreateBonExchangeTable extends Migration
             $table->bigIncrements('id');
             $table->string('description')->nullable();
             $table->unsignedBigInteger('value')->default(0);
-            $table->unsignedInteger('cost')->default(0);
+            $table->unsignedBigInteger('cost')->default(0);
             $table->boolean('upload')->default(0);
             $table->boolean('download')->default(0);
             $table->boolean('personal_freeleech')->default(0);
